Move JWT to shared
